Generate stable IDs in the HTML output
With generate.consistent.ids=1, the indexed terms throughout the book
will have stable identifiers as anchors (as long as the structure of the
text does not change). With autolink.index.see=0, no identifiers will be
generated internally for the index entries.

As a result, rerunning the HTML conversion should not change any
identifier, keeping the HTML files comparable across versions.